requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
chromadb==0.5.0
sentence-transformers==3.0.1
httpx==0.27.0
google-generativeai==0.7.2
python-multipart==0.0.9
